logging-service/requirements.txt
fastapi
uvicorn
pydantic
hazelcast-python-client
python-consul